+++ inst/include/jmotif.h
@@ -91,7 +91,8 @@ public:
   ~VisitRegistry();
 };
 //
-Rcpp::DataFrame find_discords_brute_force(NumericVector ts, int w_size, int discords_num);
+Rcpp::DataFrame find_discords_brute_force(NumericVector ts, int w_size, int discords_num,
+                                          double n_threshold);
 //
 
 //
@@ -269,6 +270,8 @@ std::vector<double> _znorm(const std::vector<double>& ts, double threshold);
 std::vector<double> _paa(std::vector<double> ts, int paa_num); // deprecated
 std::vector<double> _paa2(std::vector<double> ts, int paa_num);
 double _euclidean_dist(std::vector<double>* seq1, std::vector<double>* seq2);
+double _early_abandoned_dist(std::vector<double>* seq1, std::vector<double>* seq2,
+                             double upper_limit);
 std::vector<double> _subseries(std::vector<double>* ts, int start, int end);
 //
 std::string _series_to_string(std::vector<double> ts, int a_size);